client side responding to socket.on within a script tag directly in the body of the index.html
