+++ npm-debug.log
@@ -1,0 +1,35 @@
+0 info it worked if it ends with ok
+1 verbose cli [ 'node', '/usr/local/bin/npm', 'start' ]
+2 info using npm@2.1.14
+3 info using node@v0.10.33
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info prestart nodeSynth@0.0.1
+6 info start nodeSynth@0.0.1
+7 verbose unsafe-perm in lifecycle true
+8 info nodeSynth@0.0.1 Failed to exec start script
+9 verbose stack Error: nodeSynth@0.0.1 start: `node index.js`
+9 verbose stack Exit status 8
+9 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:212:16)
+9 verbose stack     at EventEmitter.emit (events.js:98:17)
+9 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:14:12)
+9 verbose stack     at ChildProcess.emit (events.js:98:17)
+9 verbose stack     at maybeClose (child_process.js:756:16)
+9 verbose stack     at Process.ChildProcess._handle.onexit (child_process.js:823:5)
+10 verbose pkgid nodeSynth@0.0.1
+11 verbose cwd /Users/aaron/repos/nodesocketsynthproject
+12 error Darwin 14.0.0
+13 error argv "node" "/usr/local/bin/npm" "start"
+14 error node v0.10.33
+15 error npm  v2.1.14
+16 error code ELIFECYCLE
+17 error nodeSynth@0.0.1 start: `node index.js`
+17 error Exit status 8
+18 error Failed at the nodeSynth@0.0.1 start script 'node index.js'.
+18 error This is most likely a problem with the nodeSynth package,
+18 error not with npm itself.
+18 error Tell the author that this fails on your system:
+18 error     node index.js
+18 error You can get their info via:
+18 error     npm owner ls nodeSynth
+18 error There is likely additional logging output above.
+19 verbose exit [ 1, true ]
